fix: 将updated_at字段转换类型从INTEGER改为BIGINT以避免2038年问题

+++ src/storage/database/task_manager.py
@@ -4,7 +4,7 @@ from datetime import datetime
 from typing import Optional, List
 from pydantic import BaseModel, Field
 from sqlalchemy.orm import Session
-from sqlalchemy import cast, Integer
+from sqlalchemy import cast, BigInteger
 
 from storage.database.shared.model import Tasks, Users
 import time
@@ -136,14 +136,14 @@ class TaskManager:
         if start_date:
             # 将 start_date 转换为毫秒时间戳
             start_timestamp = int(start_date.timestamp() * 1000)
-            # 将 updated_at 从字符串转换为整数后再比较
-            query = query.filter(cast(Tasks.updated_at, Integer) >= start_timestamp)
+            # 将 updated_at 从字符串转换为 bigint 后再比较
+            query = query.filter(cast(Tasks.updated_at, BigInteger) >= start_timestamp)
 
         if end_date:
             # 将 end_date 转换为毫秒时间戳
             end_timestamp = int(end_date.timestamp() * 1000)
-            # 将 updated_at 从字符串转换为整数后再比较
-            query = query.filter(cast(Tasks.updated_at, Integer) < end_timestamp)
+            # 将 updated_at 从字符串转换为 bigint 后再比较
+            query = query.filter(cast(Tasks.updated_at, BigInteger) < end_timestamp)
 
         for attr, value in filters.items():
             if hasattr(Tasks, attr):
